google-api-service registered as eureka client
