+++ eureka/src/main/java/com/bdonor/eureka/eureka/EurekaApplication.java
@@ -22,15 +22,3 @@ public class EurekaApplication {
 
 }
 
-@RestController
-class ServiceInstanceRestController {
-
-	@Autowired
-	private DiscoveryClient discoveryClient;
-
-	@RequestMapping("/service-instances/{applicationName}")
-	public List<ServiceInstance> serviceInstancesByApplicationName(
-			@PathVariable String applicationName) {
-		return this.discoveryClient.getInstances(applicationName);
-	}
-}
