+++ eureka/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.port=8001
 
-eureka.client.register-with-eureka=true
-eureka.client.fetch-registry=true
+eureka.client.register-with-eureka=false
+eureka.client.fetch-registry=false
 
 logging.level.com.netflix.eureka=OFF
 logging.level.com.netflix.discovery=OFF
