+++ google-api-service/application.yml
@@ -1,0 +1,9 @@
+spring:
+  application:
+    name: google-api-service
+eureka:
+  client:
+    serviceUrl:
+      defaultZone: ${EUREKA_URI:http://localhost:8001/eureka}
+  instance:
+    preferIpAddress: true
