+++ google-api-service/src/main/java/com/bdonor/googleapiservice/GoogleApiServiceApplication.java
@@ -1,13 +1,46 @@
 package com.bdonor.googleapiservice;
 
+import com.netflix.discovery.EurekaClient;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.cloud.client.discovery.DiscoveryClient;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @SpringBootApplication
+@RestController
+@EnableDiscoveryClient
 public class GoogleApiServiceApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(GoogleApiServiceApplication.class, args);
 	}
 
+	/*
+	@Override
+	public String greeting() {
+		return String.format("Hello from '%s'!", eurekaClient.getApplication(appName).getName());
+	}*/
+}
+
+@RestController
+class ServiceInstanceRestController {
+
+	@Autowired
+	private DiscoveryClient discoveryClient;
+
+	@RequestMapping("/service-instances/{applicationName}")
+	public List<ServiceInstance> serviceInstancesByApplicationName(
+			@PathVariable String applicationName) {
+		return this.discoveryClient.getInstances(applicationName);
+	}
 }
