+++ google-api-service/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+server.port=8110
